fix naming from client to agent
